refactor: Organised namespaces
