+++ src/Dfe.PlanTech.Infrastructure.Contentful/Content/Renderers/Interfaces/IRichTextContentPartRendererCollection.cs
@@ -1,7 +1,6 @@
 using Dfe.PlanTech.Domain.Content.Interfaces;
-using Dfe.PlanTech.Infrastructure.Contentful.Content.Renderers;
 
-namespace Dfe.PlanTech.Infrastructure.Contentful.Content;
+namespace Dfe.PlanTech.Infrastructure.Contentful.Content.Renderers.Interfaces;
 
 public interface IRichTextContentPartRendererCollection
 {
