+++ src/Dfe.PlanTech.Infrastructure.Contentful/Content/Renderers/Interfaces/IRichTextRenderer.cs
@@ -1,8 +1,8 @@
 using Dfe.PlanTech.Domain.Content.Interfaces;
 
-namespace Dfe.PlanTech.Infrastructure.Contentful.Content;
+namespace Dfe.PlanTech.Infrastructure.Contentful.Content.Renderers.Interfaces;
 
 public interface IRichTextRenderer
-{    
+{
     public string ToHtml(IRichTextContent content);
 }
