+++ src/Dfe.PlanTech.Infrastructure.Contentful/Content/Renderers/Models/PartRenderers/ListItemRenderer.cs
@@ -2,22 +2,23 @@
 using System.Text;
 using Dfe.PlanTech.Domain.Content.Interfaces;
 using Dfe.PlanTech.Domain.Content.Models;
+using Dfe.PlanTech.Infrastructure.Contentful.Content.Renderers.Interfaces;
 
-namespace Dfe.PlanTech.Infrastructure.Contentful.Content.Renderers.PartRenderers;
+namespace Dfe.PlanTech.Infrastructure.Contentful.Content.Renderers.Models.PartRenderers;
 
-public class UnorderedListRenderer : RichTextContentRender
+public class ListItemRenderer : RichTextContentRender
 {
-    public UnorderedListRenderer() : base(RichTextNodeType.UnorderedList)
+    public ListItemRenderer() : base(RichTextNodeType.ListItem)
     {
     }
 
     public override StringBuilder AddHtml(IRichTextContent content, IRichTextContentPartRendererCollection renderers, StringBuilder stringBuilder)
     {
-        stringBuilder.Append("<ul>");
+        stringBuilder.Append("<li>");
 
         RenderChildren(content, renderers, stringBuilder);
 
-        stringBuilder.Append("</ul>");
+        stringBuilder.Append("</li>");
 
         return stringBuilder;
     }
