+++ src/Dfe.PlanTech.Infrastructure.Contentful/Content/Renderers/Models/PartRenderers/OrderedListRenderer.cs
@@ -2,22 +2,23 @@
 using System.Text;
 using Dfe.PlanTech.Domain.Content.Interfaces;
 using Dfe.PlanTech.Domain.Content.Models;
+using Dfe.PlanTech.Infrastructure.Contentful.Content.Renderers.Interfaces;
 
-namespace Dfe.PlanTech.Infrastructure.Contentful.Content.Renderers.PartRenderers;
+namespace Dfe.PlanTech.Infrastructure.Contentful.Content.Renderers.Models.PartRenderers;
 
-public class ListItemRenderer : RichTextContentRender
+public class OrderedListRenderer : RichTextContentRender
 {
-    public ListItemRenderer() : base(RichTextNodeType.ListItem)
+    public OrderedListRenderer() : base(RichTextNodeType.OrderedList)
     {
     }
 
     public override StringBuilder AddHtml(IRichTextContent content, IRichTextContentPartRendererCollection renderers, StringBuilder stringBuilder)
     {
-        stringBuilder.Append("<li>");
+        stringBuilder.Append("<ol>");
 
         RenderChildren(content, renderers, stringBuilder);
 
-        stringBuilder.Append("</li>");
+        stringBuilder.Append("</ol>");
 
         return stringBuilder;
     }
