+++ src/Dfe.PlanTech.Infrastructure.Contentful/Content/Renderers/Models/PartRenderers/ParagraphRenderer.cs
@@ -1,9 +1,10 @@
 using System.Text;
 using Dfe.PlanTech.Domain.Content.Interfaces;
 using Dfe.PlanTech.Domain.Content.Models;
+using Dfe.PlanTech.Infrastructure.Contentful.Content.Renderers.Interfaces;
 using Dfe.PlanTech.Infrastructure.Contentful.Content.Renderers.Options;
 
-namespace Dfe.PlanTech.Infrastructure.Contentful.Content;
+namespace Dfe.PlanTech.Infrastructure.Contentful.Content.Renderers.Models.PartRenderers;
 
 public class ParagraphRenderer : RichTextContentRender
 {
