+++ src/Dfe.PlanTech.Infrastructure.Contentful/Content/Renderers/Models/PartRenderers/TextRenderer.cs
@@ -2,8 +2,9 @@
 using System.Text;
 using Dfe.PlanTech.Domain.Content.Interfaces;
 using Dfe.PlanTech.Domain.Content.Models;
+using Dfe.PlanTech.Infrastructure.Contentful.Content.Renderers.Interfaces;
 
-namespace Dfe.PlanTech.Infrastructure.Contentful.Content.Renderers.PartRenderers;
+namespace Dfe.PlanTech.Infrastructure.Contentful.Content.Renderers.Models.PartRenderers;
 
 public class TextRenderer : RichTextContentRender
 {
