+++ src/Dfe.PlanTech.Infrastructure.Contentful/Content/Renderers/Models/RichTextRenderer.cs
@@ -1,8 +1,8 @@
 using System.Text;
 using Dfe.PlanTech.Domain.Content.Interfaces;
-using Dfe.PlanTech.Infrastructure.Contentful.Content.Renderers;
+using Dfe.PlanTech.Infrastructure.Contentful.Content.Renderers.Interfaces;
 
-namespace Dfe.PlanTech.Infrastructure.Contentful.Content;
+namespace Dfe.PlanTech.Infrastructure.Contentful.Content.Renderers.Models;
 
 public class RichTextRenderer : IRichTextRenderer, IRichTextContentPartRendererCollection
 {
